test(mitm-addon): split model provider response usage tests
