--- crates/runner/mitm-addon/tests/test_model_provider_response_usage.py
@@ -1,1364 +1,0 @@
-"""Tests for model-provider response usage reporting paths."""
-
-import gzip
-import json
-import zlib
-from collections.abc import Callable
-from dataclasses import dataclass
-from pathlib import Path
-
-import brotli
-import pytest
-import zstandard
-from mitmproxy.test import tutils
-
-import flow_metadata_keys as metadata_keys
-import logging_utils
-import mitm_addon
-import usage
-from body_limits import STREAM_BUFFER_LIMIT, STREAM_DECODE_CHUNK_LIMIT
-from tests.flow_helpers import header_map, response_stream
-from tests.jsonl_log_helpers import (
-    jsonl_exists_after_flush,
-    read_jsonl_entries_after_flush,
-    read_jsonl_text_after_flush,
-)
-from tests.stream_buffer_helpers import set_response_stream_buffer
-
-
-@dataclass(frozen=True)
-class ModelProviderJsonCase:
-    id: str
-    host: str
-    original_url: str
-    firewall_name: str
-    cli_agent_type: str | None
-    message_id: str
-    model: str
-    uses_openai_responses: bool
-    input_tokens: int = 50
-    output_tokens: int = 200
-    cached_tokens: int | None = None
-
-
-@dataclass(frozen=True)
-class JsonCompressionFailureCase:
-    id: str
-    make_body: Callable[[bytes], bytes]
-    content_encoding: str
-    expected_error: str
-
-
-ANTHROPIC_JSON_CASE = ModelProviderJsonCase(
-    id="anthropic",
-    host="api.anthropic.com",
-    original_url="https://api.anthropic.com/v1/messages",
-    firewall_name="model-provider:anthropic-api-key",
-    cli_agent_type=None,
-    message_id="msg_1",
-    model="claude-sonnet-4-6",
-    uses_openai_responses=False,
-)
-
-OPENAI_RESPONSES_CASE = ModelProviderJsonCase(
-    id="openai",
-    host="api.openai.com",
-    original_url="https://api.openai.com/v1/responses",
-    firewall_name="model-provider:openai-api-key",
-    cli_agent_type="codex",
-    message_id="resp_1",
-    model="gpt-5.5",
-    uses_openai_responses=True,
-    cached_tokens=10,
-)
-
-CODEX_OAUTH_RESPONSES_CASE = ModelProviderJsonCase(
-    id="codex-oauth",
-    host="chatgpt.com",
-    original_url="https://chatgpt.com/backend-api/codex/responses",
-    firewall_name="model-provider:codex-oauth-token",
-    cli_agent_type="codex",
-    message_id="resp_1",
-    model="gpt-5.5",
-    uses_openai_responses=True,
-    cached_tokens=10,
-)
-
-MODEL_PROVIDER_JSON_CASES = (ANTHROPIC_JSON_CASE, OPENAI_RESPONSES_CASE)
-
-
-def _model_provider_json_case_id(provider_case: ModelProviderJsonCase) -> str:
-    return provider_case.id
-
-
-def _json_compression_failure_case_id(encoding_case: JsonCompressionFailureCase) -> str:
-    return encoding_case.id
-
-
-def _identity_body(payload: bytes) -> bytes:
-    return payload
-
-
-def _raw_deflate_body(payload: bytes) -> bytes:
-    compressor = zlib.compressobj(wbits=-zlib.MAX_WBITS)
-    return compressor.compress(payload) + compressor.flush()
-
-
-def _truncated_gzip_prefix(payload: bytes) -> bytes:
-    return gzip.compress(payload)[:10]
-
-
-def _truncated_gzip_trailer(payload: bytes) -> bytes:
-    return gzip.compress(payload)[:-1]
-
-
-def _truncated_deflate_trailer(payload: bytes) -> bytes:
-    return zlib.compress(payload)[:-1]
-
-
-def _empty_gzip_member_before_garbage(_payload: bytes) -> bytes:
-    return gzip.compress(b"") + b"garbage"
-
-
-def _empty_deflate_stream_before_garbage(_payload: bytes) -> bytes:
-    return zlib.compress(b"") + b"garbage"
-
-
-def _truncated_brotli_prefix(payload: bytes) -> bytes:
-    return brotli.compress(payload)[:2]
-
-
-def _truncated_brotli_trailer(payload: bytes) -> bytes:
-    return brotli.compress(payload)[:-1]
-
-
-def _truncated_zstd_prefix(payload: bytes) -> bytes:
-    return zstandard.ZstdCompressor().compress(payload)[:5]
-
-
-def _zstd_frame_before_garbage(payload: bytes) -> bytes:
-    return zstandard.ZstdCompressor().compress(payload) + b"garbage"
-
-
-def _zstd_frame_before_truncated_frame(payload: bytes) -> bytes:
-    trailing_frame = zstandard.ZstdCompressor().compress(b"{}")
-    return zstandard.ZstdCompressor().compress(payload) + trailing_frame[:5]
-
-
-JSON_COMPRESSION_FAILURE_CASES = (
-    JsonCompressionFailureCase(
-        id="chained-gzip",
-        make_body=gzip.compress,
-        content_encoding="gzip, identity",
-        expected_error="unsupported content encoding",
-    ),
-    JsonCompressionFailureCase(
-        id="raw-json-with-unknown-header",
-        make_body=_identity_body,
-        content_encoding="x-custom",
-        expected_error="unsupported content encoding",
-    ),
-    JsonCompressionFailureCase(
-        id="raw-deflate",
-        make_body=_raw_deflate_body,
-        content_encoding="deflate",
-        expected_error="invalid compressed body",
-    ),
-    JsonCompressionFailureCase(
-        id="raw-json-with-gzip-header",
-        make_body=_identity_body,
-        content_encoding="gzip",
-        expected_error="invalid compressed body",
-    ),
-    JsonCompressionFailureCase(
-        id="raw-json-with-br-header",
-        make_body=_identity_body,
-        content_encoding="br",
-        expected_error="invalid compressed body",
-    ),
-    JsonCompressionFailureCase(
-        id="raw-json-with-zstd-header",
-        make_body=_identity_body,
-        content_encoding="zstd",
-        expected_error="invalid compressed body",
-    ),
-    JsonCompressionFailureCase(
-        id="truncated-gzip-prefix",
-        make_body=_truncated_gzip_prefix,
-        content_encoding="gzip",
-        expected_error="incomplete compressed body",
-    ),
-    JsonCompressionFailureCase(
-        id="truncated-gzip-trailer",
-        make_body=_truncated_gzip_trailer,
-        content_encoding="gzip",
-        expected_error="incomplete compressed body",
-    ),
-    JsonCompressionFailureCase(
-        id="truncated-deflate-trailer",
-        make_body=_truncated_deflate_trailer,
-        content_encoding="deflate",
-        expected_error="incomplete compressed body",
-    ),
-    JsonCompressionFailureCase(
-        id="empty-gzip-member-before-garbage",
-        make_body=_empty_gzip_member_before_garbage,
-        content_encoding="gzip",
-        expected_error="invalid compressed body",
-    ),
-    JsonCompressionFailureCase(
-        id="empty-deflate-stream-before-garbage",
-        make_body=_empty_deflate_stream_before_garbage,
-        content_encoding="deflate",
-        expected_error="invalid compressed body",
-    ),
-    JsonCompressionFailureCase(
-        id="truncated-brotli-prefix",
-        make_body=_truncated_brotli_prefix,
-        content_encoding="br",
-        expected_error="incomplete compressed body",
-    ),
-    JsonCompressionFailureCase(
-        id="truncated-brotli-trailer",
-        make_body=_truncated_brotli_trailer,
-        content_encoding="br",
-        expected_error="incomplete compressed body",
-    ),
-    JsonCompressionFailureCase(
-        id="truncated-zstd-prefix",
-        make_body=_truncated_zstd_prefix,
-        content_encoding="zstd",
-        expected_error="incomplete compressed body",
-    ),
-    JsonCompressionFailureCase(
-        id="zstd-frame-before-garbage",
-        make_body=_zstd_frame_before_garbage,
-        content_encoding="zstd",
-        expected_error="invalid compressed body",
-    ),
-    JsonCompressionFailureCase(
-        id="zstd-frame-before-truncated-frame",
-        make_body=_zstd_frame_before_truncated_frame,
-        content_encoding="zstd",
-        expected_error="incomplete compressed body",
-    ),
-)
-
-
-def _standard_success_payload(
-    provider_case: ModelProviderJsonCase,
-    *,
-    input_tokens: int | None = None,
-    output_tokens: int | None = None,
-    cached_tokens: int | None = None,
-) -> bytes:
-    resolved_input_tokens = provider_case.input_tokens if input_tokens is None else input_tokens
-    resolved_output_tokens = provider_case.output_tokens if output_tokens is None else output_tokens
-    resolved_cached_tokens = provider_case.cached_tokens if cached_tokens is None else cached_tokens
-    usage_payload: dict[str, object] = {
-        "input_tokens": resolved_input_tokens,
-        "output_tokens": resolved_output_tokens,
-    }
-    if provider_case.uses_openai_responses and resolved_cached_tokens is not None:
-        usage_payload["input_tokens_details"] = {
-            "cached_tokens": resolved_cached_tokens,
-        }
-    payload: dict[str, object] = {
-        "id": provider_case.message_id,
-        "model": provider_case.model,
-        "usage": usage_payload,
-    }
-    if not provider_case.uses_openai_responses:
-        payload["content"] = [{"type": "text", "text": "Hello"}]
-    return json.dumps(payload).encode()
-
-
-def _expected_usage(
-    provider_case: ModelProviderJsonCase,
-    *,
-    input_tokens: int | None = None,
-    output_tokens: int | None = None,
-    cached_tokens: int | None = None,
-) -> dict[str, object]:
-    resolved_input_tokens = provider_case.input_tokens if input_tokens is None else input_tokens
-    resolved_output_tokens = provider_case.output_tokens if output_tokens is None else output_tokens
-    resolved_cached_tokens = provider_case.cached_tokens if cached_tokens is None else cached_tokens
-    expected = {
-        "message_id": provider_case.message_id,
-        "model": provider_case.model,
-        "tokens.input": resolved_input_tokens,
-        "tokens.output": resolved_output_tokens,
-    }
-    if provider_case.uses_openai_responses and resolved_cached_tokens is not None:
-        assert resolved_cached_tokens <= resolved_input_tokens
-        expected["tokens.input"] = resolved_input_tokens - resolved_cached_tokens
-        expected["tokens.cache_read"] = resolved_cached_tokens
-    return expected
-
-
-def _expected_event_quantities(provider_case: ModelProviderJsonCase) -> dict[str, int]:
-    return {
-        category: quantity
-        for category, quantity in _expected_usage(provider_case).items()
-        if category.startswith("tokens.")
-        and isinstance(quantity, int)
-        and not isinstance(quantity, bool)
-        and quantity > 0
-    }
-
-
-class TestModelProviderResponseUsage:
-    """Tests for model-provider JSON usage extraction and response reporting."""
-
-    @pytest.fixture(autouse=True)
-    def _sync_usage_delivery(self, sync_usage_executor, usage_webhook_api):
-        self._usage_webhook_api = usage_webhook_api
-
-    def _run_response(self, flow):
-        with self._usage_webhook_api() as webhook:
-            mitm_addon.response(flow)
-            usage.flush_usage_events(trigger="test")
-        return webhook
-
-    def _set_common_model_metadata(
-        self,
-        flow,
-        tmp_path: Path,
-        *,
-        billable: bool = True,
-        proxy_log_path: Path | None = None,
-        run_id: str = "run-abc-123",
-    ) -> None:
-        flow.metadata[metadata_keys.VM_RUN_ID] = run_id
-        flow.metadata[metadata_keys.VM_NETWORK_LOG_PATH] = str(tmp_path / "network.jsonl")
-        if proxy_log_path is not None:
-            flow.metadata[metadata_keys.VM_PROXY_LOG_PATH] = str(proxy_log_path)
-        flow.metadata[metadata_keys.FIREWALL_ACTION] = "ALLOW"
-        flow.metadata[metadata_keys.FIREWALL_BILLABLE] = billable
-        flow.metadata[metadata_keys.VM_SANDBOX_AUTH_KEY] = "tok-xyz"
-
-    def _set_model_provider_metadata(
-        self,
-        flow,
-        tmp_path: Path,
-        provider_case: ModelProviderJsonCase,
-        *,
-        billable: bool = True,
-        observable: bool = True,
-        proxy_log_path: Path | None = None,
-        run_id: str = "run-abc-123",
-    ) -> None:
-        self._set_common_model_metadata(
-            flow,
-            tmp_path,
-            billable=billable,
-            proxy_log_path=proxy_log_path,
-            run_id=run_id,
-        )
-        flow.metadata[metadata_keys.ORIGINAL_URL] = provider_case.original_url
-        flow.metadata[metadata_keys.FIREWALL_NAME] = provider_case.firewall_name
-        if observable:
-            flow.metadata[metadata_keys.MODEL_USAGE_PROVIDER] = provider_case.model
-        if provider_case.cli_agent_type is not None:
-            flow.metadata[metadata_keys.CLI_AGENT_TYPE] = provider_case.cli_agent_type
-
-    def _model_provider_flow(
-        self,
-        real_flow,
-        tmp_path: Path,
-        provider_case: ModelProviderJsonCase,
-        *,
-        billable: bool = True,
-        observable: bool = True,
-        proxy_log_path: Path | None = None,
-        run_id: str = "run-abc-123",
-    ):
-        flow = real_flow(with_response=False, host=provider_case.host)
-        self._set_model_provider_metadata(
-            flow,
-            tmp_path,
-            provider_case,
-            billable=billable,
-            observable=observable,
-            proxy_log_path=proxy_log_path,
-            run_id=run_id,
-        )
-        return flow
-
-    def test_non_streaming_json_fallback(self, tmp_path, real_flow):
-        """Non-streaming JSON response should extract usage from buffer."""
-        provider_case = ANTHROPIC_JSON_CASE
-        flow = self._model_provider_flow(real_flow, tmp_path, provider_case)
-        # No model_provider_usage set (no SSE parser) — JSON body in buffer
-        body = _standard_success_payload(provider_case)
-        set_response_stream_buffer(flow, body)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map(
-                {"content-type": "application/json", "content-length": str(len(body))}
-            ),
-        )
-
-        self._run_response(flow)
-
-        # JSON fallback should populate model_provider_usage in metadata
-        extracted = flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]
-        expected = _expected_usage(provider_case)
-        assert extracted["model"] == expected["model"]
-        assert extracted["tokens.input"] == expected["tokens.input"]
-        assert extracted["tokens.output"] == expected["tokens.output"]
-
-    def test_openai_non_streaming_json_fallback(self, tmp_path, real_flow):
-        """Legacy JSON fallback should use OpenAI Responses mapping."""
-        provider_case = OPENAI_RESPONSES_CASE
-        flow = self._model_provider_flow(real_flow, tmp_path, provider_case)
-        body = _standard_success_payload(provider_case)
-        set_response_stream_buffer(flow, body)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map(
-                {"content-type": "application/json", "content-length": str(len(body))}
-            ),
-        )
-
-        webhook = self._run_response(flow)
-
-        extracted = flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]
-        expected = _expected_usage(provider_case)
-        assert extracted["message_id"] == expected["message_id"]
-        assert extracted["model"] == expected["model"]
-        assert extracted["tokens.input"] == expected["tokens.input"]
-        assert extracted["tokens.output"] == expected["tokens.output"]
-        assert extracted["tokens.cache_read"] == expected["tokens.cache_read"]
-        events = webhook.usage_events()
-        by_category = {event["category"]: event["quantity"] for event in events}
-        assert by_category == _expected_event_quantities(provider_case)
-
-    def test_anthropic_json_fallback_parse_error_logs_proxy_warning(self, tmp_path, real_flow):
-        """Legacy JSON fallback parse failures should be observable."""
-        proxy_log_path = tmp_path / "proxy.jsonl"
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            ANTHROPIC_JSON_CASE,
-            proxy_log_path=proxy_log_path,
-        )
-        body = b'{"id":"msg_1","model":"claude-sonnet-4-6","usage":{"input_tokens":50'
-        set_response_stream_buffer(flow, body)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "application/json"}),
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        assert metadata_keys.MODEL_PROVIDER_USAGE not in flow.metadata
-        entries = read_jsonl_entries_after_flush(proxy_log_path)
-        assert len(entries) == 1
-        assert entries[0]["level"] == "warn"
-        assert entries[0]["message"] == "Model provider JSON usage extraction failed"
-        assert entries[0]["type"] == "usage_event"
-        assert entries[0]["error"] == "incomplete json"
-
-    def test_json_fallback_parser_bound_error_logs_proxy_warning(self, tmp_path, real_flow):
-        """Bounded parser failures should be observable without logging body content."""
-        proxy_log_path = tmp_path / "proxy.jsonl"
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            ANTHROPIC_JSON_CASE,
-            proxy_log_path=proxy_log_path,
-        )
-        oversized_model = "x" * 1025
-        body = json.dumps(
-            {
-                "id": "msg_1",
-                "model": oversized_model,
-                "usage": {"input_tokens": 50},
-            }
-        ).encode()
-        set_response_stream_buffer(flow, body)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "application/json"}),
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        assert metadata_keys.MODEL_PROVIDER_USAGE not in flow.metadata
-        entries = read_jsonl_entries_after_flush(proxy_log_path)
-        assert len(entries) == 1
-        assert entries[0]["level"] == "warn"
-        assert entries[0]["message"] == "Model provider JSON usage extraction failed"
-        assert entries[0]["type"] == "usage_event"
-        assert entries[0]["error"] == "string limit exceeded"
-        assert oversized_model not in read_jsonl_text_after_flush(proxy_log_path)
-
-    def test_openai_json_fallback_parse_error_logs_proxy_warning(self, tmp_path, real_flow):
-        """OpenAI fallback parse failures should use the same proxy warning."""
-        proxy_log_path = tmp_path / "proxy.jsonl"
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            OPENAI_RESPONSES_CASE,
-            proxy_log_path=proxy_log_path,
-        )
-        body = b'{"id":"resp_1","model":"gpt-5.5","usage":{"input_tokens":50'
-        set_response_stream_buffer(flow, body)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "application/json"}),
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        assert metadata_keys.MODEL_PROVIDER_USAGE not in flow.metadata
-        entries = read_jsonl_entries_after_flush(proxy_log_path)
-        assert len(entries) == 1
-        assert entries[0]["level"] == "warn"
-        assert entries[0]["message"] == "Model provider JSON usage extraction failed"
-        assert entries[0]["type"] == "usage_event"
-        assert entries[0]["error"] == "incomplete json"
-
-    @pytest.mark.parametrize(
-        "encoding_case",
-        JSON_COMPRESSION_FAILURE_CASES,
-        ids=_json_compression_failure_case_id,
-    )
-    @pytest.mark.parametrize(
-        "provider_case",
-        MODEL_PROVIDER_JSON_CASES,
-        ids=_model_provider_json_case_id,
-    )
-    def test_json_fallback_compressed_body_parse_failure_logs_proxy_warning(
-        self,
-        tmp_path,
-        real_flow,
-        encoding_case,
-        provider_case,
-    ):
-        """One-shot decompression failures leave compressed bytes and log parse failure."""
-        proxy_log_path = tmp_path / "proxy.jsonl"
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            provider_case,
-            proxy_log_path=proxy_log_path,
-        )
-        payload = _standard_success_payload(provider_case)
-        body = encoding_case.make_body(payload)
-
-        set_response_stream_buffer(flow, body)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map(
-                {
-                    "content-type": "application/json",
-                    "content-encoding": encoding_case.content_encoding,
-                }
-            ),
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        assert metadata_keys.MODEL_PROVIDER_USAGE not in flow.metadata
-        entries = read_jsonl_entries_after_flush(proxy_log_path)
-        assert len(entries) == 1
-        assert entries[0]["level"] == "warn"
-        assert entries[0]["message"] == "Model provider JSON usage extraction failed"
-        assert entries[0]["type"] == "usage_event"
-        assert entries[0]["error"] == encoding_case.expected_error
-
-    @pytest.mark.parametrize("encoding_case", ["gzip", "deflate"])
-    @pytest.mark.parametrize(
-        "provider_case",
-        MODEL_PROVIDER_JSON_CASES,
-        ids=_model_provider_json_case_id,
-    )
-    def test_json_fallback_concatenated_zlib_member_reports_usage(
-        self,
-        tmp_path,
-        real_flow,
-        encoding_case,
-        provider_case,
-    ):
-        """Zlib stream concatenation should not let an empty first member hide usage."""
-        proxy_log_path = tmp_path / "proxy.jsonl"
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            provider_case,
-            proxy_log_path=proxy_log_path,
-        )
-        payload = _standard_success_payload(provider_case)
-        if encoding_case == "gzip":
-            body = gzip.compress(b"") + gzip.compress(payload)
-        else:
-            body = zlib.compress(b"") + zlib.compress(payload)
-        set_response_stream_buffer(flow, body)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map(
-                {
-                    "content-type": "application/json",
-                    "content-encoding": encoding_case,
-                }
-            ),
-        )
-
-        self._run_response(flow)
-
-        extracted = flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]
-        expected = _expected_usage(provider_case)
-        assert extracted["model"] == expected["model"]
-        assert extracted["tokens.input"] == expected["tokens.input"]
-        assert extracted["tokens.output"] == expected["tokens.output"]
-        if provider_case.uses_openai_responses:
-            assert extracted["tokens.cache_read"] == expected["tokens.cache_read"]
-        if jsonl_exists_after_flush(proxy_log_path):
-            entries = read_jsonl_entries_after_flush(proxy_log_path)
-            assert not any(
-                entry.get("message") == "Model provider JSON usage extraction failed"
-                for entry in entries
-            )
-
-    @pytest.mark.parametrize("encoding_case", ["br", "zstd"])
-    @pytest.mark.parametrize(
-        "provider_case",
-        MODEL_PROVIDER_JSON_CASES,
-        ids=_model_provider_json_case_id,
-    )
-    def test_json_fallback_brotli_and_zstd_report_usage(
-        self,
-        tmp_path,
-        real_flow,
-        encoding_case,
-        provider_case,
-    ):
-        """Diagnostic fallback should handle complete br/zstd JSON bodies."""
-        proxy_log_path = tmp_path / "proxy.jsonl"
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            provider_case,
-            proxy_log_path=proxy_log_path,
-        )
-        payload = _standard_success_payload(provider_case)
-
-        if encoding_case == "br":
-            body = brotli.compress(payload)
-        else:
-            body = zstandard.ZstdCompressor().compress(payload)
-
-        set_response_stream_buffer(flow, body)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map(
-                {
-                    "content-type": "application/json",
-                    "content-encoding": encoding_case,
-                }
-            ),
-        )
-
-        webhook = self._run_response(flow)
-
-        extracted = flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]
-        expected_usage = _expected_usage(provider_case)
-        assert extracted["model"] == expected_usage["model"]
-        assert extracted["tokens.input"] == expected_usage["tokens.input"]
-        assert extracted["tokens.output"] == expected_usage["tokens.output"]
-        if provider_case.uses_openai_responses:
-            assert extracted["tokens.cache_read"] == expected_usage["tokens.cache_read"]
-        events = webhook.usage_events()
-        by_category = {event["category"]: event["quantity"] for event in events}
-        assert by_category == _expected_event_quantities(provider_case)
-        if jsonl_exists_after_flush(proxy_log_path):
-            entries = read_jsonl_entries_after_flush(proxy_log_path)
-            assert not any(
-                entry.get("message") == "Model provider JSON usage extraction failed"
-                for entry in entries
-            )
-
-    def test_json_fallback_valid_body_without_usage_stays_quiet(self, tmp_path, real_flow):
-        """Valid JSON without usage is not a parser failure."""
-        proxy_log_path = tmp_path / "proxy.jsonl"
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            ANTHROPIC_JSON_CASE,
-            proxy_log_path=proxy_log_path,
-        )
-        body = b'{"id":"msg_1"}'
-        set_response_stream_buffer(flow, body)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "application/json"}),
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        assert metadata_keys.MODEL_PROVIDER_USAGE not in flow.metadata
-        logging_utils.flush_log_path(str(proxy_log_path))
-        assert not proxy_log_path.exists()
-
-    def test_openai_json_fallback_valid_body_without_usage_stays_quiet(self, tmp_path, real_flow):
-        """OpenAI fallback should also keep valid no-usage JSON quiet."""
-        proxy_log_path = tmp_path / "proxy.jsonl"
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            OPENAI_RESPONSES_CASE,
-            proxy_log_path=proxy_log_path,
-        )
-        body = b'{"id":"resp_1","model":"gpt-5.5"}'
-        set_response_stream_buffer(flow, body)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "application/json"}),
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        assert metadata_keys.MODEL_PROVIDER_USAGE not in flow.metadata
-        logging_utils.flush_log_path(str(proxy_log_path))
-        assert not proxy_log_path.exists()
-
-    @pytest.mark.parametrize(
-        "encoding_case", ["identity", "gzip", "deflate", "br", "zstd", "zstd-no-size"]
-    )
-    @pytest.mark.parametrize(
-        "provider_case",
-        MODEL_PROVIDER_JSON_CASES,
-        ids=_model_provider_json_case_id,
-    )
-    def test_json_fallback_empty_body_stays_quiet(
-        self,
-        tmp_path,
-        real_flow,
-        encoding_case,
-        provider_case,
-    ):
-        """Empty model-provider bodies are not JSON parser failures."""
-        proxy_log_path = tmp_path / "proxy.jsonl"
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            provider_case,
-            proxy_log_path=proxy_log_path,
-        )
-        if encoding_case == "gzip":
-            body = gzip.compress(b"")
-        elif encoding_case == "deflate":
-            body = zlib.compress(b"")
-        elif encoding_case == "br":
-            body = brotli.compress(b"")
-        elif encoding_case == "zstd":
-            body = zstandard.ZstdCompressor().compress(b"")
-        elif encoding_case == "zstd-no-size":
-            body = zstandard.ZstdCompressor(write_content_size=False).compress(b"")
-        else:
-            body = b""
-        response_headers = {
-            "content-type": "application/json",
-            "content-length": str(len(body)),
-        }
-        if encoding_case != "identity":
-            response_headers["content-encoding"] = (
-                "zstd" if encoding_case == "zstd-no-size" else encoding_case
-            )
-        set_response_stream_buffer(flow, body)
-        flow.response = tutils.tresp(status_code=200, headers=header_map(response_headers))
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        assert metadata_keys.MODEL_PROVIDER_USAGE not in flow.metadata
-        logging_utils.flush_log_path(str(proxy_log_path))
-        assert not proxy_log_path.exists()
-
-    def test_anthropic_json_fallback_metadata_only_usage_stays_quiet(self, tmp_path, real_flow):
-        """Anthropic metadata without positive token usage is not a parser failure."""
-        proxy_log_path = tmp_path / "proxy.jsonl"
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            ANTHROPIC_JSON_CASE,
-            proxy_log_path=proxy_log_path,
-        )
-        body = b'{"id":"msg_1","model":"claude-sonnet-4-6"}'
-        set_response_stream_buffer(flow, body)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "application/json"}),
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE] == {
-            "message_id": "msg_1",
-            "model": "claude-sonnet-4-6",
-        }
-        logging_utils.flush_log_path(str(proxy_log_path))
-        assert not proxy_log_path.exists()
-
-    @pytest.mark.parametrize(
-        "provider_case",
-        MODEL_PROVIDER_JSON_CASES,
-        ids=_model_provider_json_case_id,
-    )
-    def test_json_fallback_zero_token_usage_stays_quiet(self, tmp_path, real_flow, provider_case):
-        """Valid zero-token usage is not a parser failure and does not bill."""
-        proxy_log_path = tmp_path / "proxy.jsonl"
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            provider_case,
-            proxy_log_path=proxy_log_path,
-        )
-        body = _standard_success_payload(
-            provider_case,
-            input_tokens=0,
-            output_tokens=0,
-            cached_tokens=0,
-        )
-        expected_usage = _expected_usage(
-            provider_case,
-            input_tokens=0,
-            output_tokens=0,
-            cached_tokens=0,
-        )
-
-        set_response_stream_buffer(flow, body)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "application/json"}),
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE] == expected_usage
-        logging_utils.flush_log_path(str(proxy_log_path))
-        assert not proxy_log_path.exists()
-
-    def test_codex_oauth_non_streaming_json_fallback(self, tmp_path, real_flow):
-        """Codex OAuth model-provider fallback uses OpenAI Responses mapping."""
-        provider_case = CODEX_OAUTH_RESPONSES_CASE
-        flow = self._model_provider_flow(real_flow, tmp_path, provider_case)
-        body = _standard_success_payload(provider_case)
-        set_response_stream_buffer(flow, body)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map(
-                {"content-type": "application/json", "content-length": str(len(body))}
-            ),
-        )
-
-        webhook = self._run_response(flow)
-
-        extracted = flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]
-        expected = _expected_usage(provider_case)
-        assert extracted["message_id"] == expected["message_id"]
-        assert extracted["model"] == expected["model"]
-        assert extracted["tokens.input"] == expected["tokens.input"]
-        assert extracted["tokens.output"] == expected["tokens.output"]
-        assert extracted["tokens.cache_read"] == expected["tokens.cache_read"]
-        events = webhook.usage_events()
-        by_category = {event["category"]: event["quantity"] for event in events}
-        assert by_category == _expected_event_quantities(provider_case)
-
-    def test_non_billable_openai_json_reports_observation_without_billing(
-        self, tmp_path, real_flow
-    ):
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            OPENAI_RESPONSES_CASE,
-            billable=False,
-        )
-        body = _standard_success_payload(OPENAI_RESPONSES_CASE)
-        set_response_stream_buffer(flow, body)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "application/json"}),
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.usage_events() == []
-        observations = webhook.model_usage_observation_events()
-        by_category = {event["category"]: event["quantity"] for event in observations}
-        assert by_category == _expected_event_quantities(OPENAI_RESPONSES_CASE)
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]["model"] == "gpt-5.5"
-
-    def test_non_observable_json_response_does_not_register_incremental_parser(
-        self,
-        tmp_path,
-        real_flow,
-    ):
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            ANTHROPIC_JSON_CASE,
-            billable=False,
-            observable=False,
-        )
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "application/json"}),
-        )
-
-        mitm_addon.responseheaders(flow)
-
-        response_stream(flow)(b"x" * (STREAM_BUFFER_LIMIT + 1000))
-
-        assert "model_json_usage_finish" not in flow.metadata
-        assert metadata_keys.MODEL_PROVIDER_USAGE not in flow.metadata
-        assert len(flow.metadata[metadata_keys.STREAM_BUFFER]) == STREAM_BUFFER_LIMIT
-        assert flow.metadata[metadata_keys.STREAM_BUFFER_STATE]["truncated"] is True
-
-    def test_non_observable_json_fallback_parse_error_stays_quiet(self, tmp_path, real_flow):
-        """Model-provider fallback without MODEL_USAGE_PROVIDER must not emit warnings."""
-        proxy_log_path = tmp_path / "proxy.jsonl"
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            OPENAI_RESPONSES_CASE,
-            billable=False,
-            observable=False,
-            proxy_log_path=proxy_log_path,
-        )
-        body = b'{"id":"resp_1","model":"gpt-5.5","usage":{"input_tokens":50'
-        set_response_stream_buffer(flow, body)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "application/json"}),
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        assert metadata_keys.MODEL_PROVIDER_USAGE not in flow.metadata
-        logging_utils.flush_log_path(str(proxy_log_path))
-        assert not proxy_log_path.exists()
-
-    def test_full_pipeline_large_model_json_uses_bounded_buffer(self, tmp_path, real_flow):
-        """responseheaders + response report model usage without full-body buffering."""
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            ANTHROPIC_JSON_CASE,
-            proxy_log_path=tmp_path / "proxy.jsonl",
-        )
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "application/json"}),
-        )
-
-        mitm_addon.responseheaders(flow)
-        callback = response_stream(flow)
-        callback(b'{"id":"msg_1","model":"claude-sonnet-4-6","content":[{"text":"')
-        callback(b"x" * (STREAM_BUFFER_LIMIT + 4096))
-        callback(b'"}],"usage":{"input_tokens":50,"output_tokens":200}}')
-        assert len(flow.metadata[metadata_keys.STREAM_BUFFER]) == STREAM_BUFFER_LIMIT
-        assert flow.metadata[metadata_keys.STREAM_BUFFER_STATE]["truncated"] is True
-
-        webhook = self._run_response(flow)
-
-        extracted = flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]
-        assert extracted["message_id"] == "msg_1"
-        assert extracted["model"] == "claude-sonnet-4-6"
-        assert extracted["tokens.input"] == 50
-        assert extracted["tokens.output"] == 200
-        events = webhook.usage_events()
-        by_category = {event["category"]: event["quantity"] for event in events}
-        assert by_category == {"tokens.input": 50, "tokens.output": 200}
-
-    @pytest.mark.parametrize(
-        "provider_case",
-        MODEL_PROVIDER_JSON_CASES,
-        ids=_model_provider_json_case_id,
-    )
-    def test_full_pipeline_compressed_model_json_reports_usage(
-        self, tmp_path, real_flow, provider_case
-    ):
-        """responseheaders parser should decompress non-SSE model JSON before extraction."""
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            provider_case,
-            proxy_log_path=tmp_path / "proxy.jsonl",
-        )
-        payload = _standard_success_payload(provider_case)
-        compressed = gzip.compress(payload)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "application/json", "content-encoding": "gzip"}),
-        )
-
-        mitm_addon.responseheaders(flow)
-        midpoint = len(compressed) // 2
-        response_stream(flow)(compressed[:midpoint])
-        response_stream(flow)(compressed[midpoint:])
-
-        webhook = self._run_response(flow)
-
-        extracted = flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]
-        expected_usage = _expected_usage(provider_case)
-        assert extracted["message_id"] == expected_usage["message_id"]
-        assert extracted["model"] == expected_usage["model"]
-        assert extracted["tokens.input"] == expected_usage["tokens.input"]
-        assert extracted["tokens.output"] == expected_usage["tokens.output"]
-        if provider_case.uses_openai_responses:
-            assert extracted["tokens.cache_read"] == expected_usage["tokens.cache_read"]
-        events = webhook.usage_events()
-        by_category = {event["category"]: event["quantity"] for event in events}
-        assert by_category == _expected_event_quantities(provider_case)
-
-    def test_full_pipeline_zstd_model_json_scans_past_decode_chunk_limit(
-        self,
-        tmp_path,
-        real_flow,
-    ):
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            ANTHROPIC_JSON_CASE,
-            proxy_log_path=tmp_path / "proxy.jsonl",
-        )
-        payload = (
-            b'{"id":"msg_zstd","model":"claude-sonnet-4-6","content":[{"text":"'
-            + b"A" * (STREAM_DECODE_CHUNK_LIMIT * 3)
-            + b'"}],"usage":{"input_tokens":10,"output_tokens":20}}'
-        )
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "application/json", "content-encoding": "zstd"}),
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(zstandard.ZstdCompressor().compress(payload))
-
-        webhook = self._run_response(flow)
-
-        extracted = flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]
-        assert extracted["message_id"] == "msg_zstd"
-        assert extracted["tokens.input"] == 10
-        assert extracted["tokens.output"] == 20
-        events = webhook.usage_events()
-        by_category = {event["category"]: event["quantity"] for event in events}
-        assert by_category == {"tokens.input": 10, "tokens.output": 20}
-
-    @pytest.mark.parametrize("encoding_case", ["gzip", "deflate"])
-    @pytest.mark.parametrize(
-        "provider_case",
-        MODEL_PROVIDER_JSON_CASES,
-        ids=_model_provider_json_case_id,
-    )
-    def test_full_pipeline_truncated_compressed_model_json_does_not_report_usage(
-        self, tmp_path, real_flow, encoding_case, provider_case
-    ):
-        """Incremental JSON usage must reject compressed streams missing a trailer."""
-        proxy_log_path = tmp_path / "proxy.jsonl"
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            provider_case,
-            proxy_log_path=proxy_log_path,
-        )
-        payload = _standard_success_payload(provider_case)
-        if encoding_case == "gzip":
-            compressed = gzip.compress(payload)[:-1]
-        else:
-            compressed = zlib.compress(payload)[:-1]
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map(
-                {"content-type": "application/json", "content-encoding": encoding_case}
-            ),
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(compressed)
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        assert metadata_keys.MODEL_PROVIDER_USAGE not in flow.metadata
-        entries = read_jsonl_entries_after_flush(proxy_log_path)
-        usage_warnings = [
-            entry
-            for entry in entries
-            if entry.get("message") == "Model provider JSON usage extraction failed"
-        ]
-        assert len(usage_warnings) == 1
-        assert usage_warnings[0]["error"] == "incomplete compressed body"
-
-    @pytest.mark.parametrize("encoding_case", ["gzip", "deflate"])
-    def test_full_pipeline_concatenated_zlib_model_json_reports_usage(
-        self, tmp_path, real_flow, encoding_case
-    ):
-        """Streaming decompression should feed later zlib members into JSON usage parsing."""
-        provider_case = ANTHROPIC_JSON_CASE
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            provider_case,
-            proxy_log_path=tmp_path / "proxy.jsonl",
-        )
-        payload = _standard_success_payload(provider_case)
-        if encoding_case == "gzip":
-            compressed = gzip.compress(b"") + gzip.compress(payload)
-        else:
-            compressed = zlib.compress(b"") + zlib.compress(payload)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map(
-                {"content-type": "application/json", "content-encoding": encoding_case}
-            ),
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(compressed)
-
-        webhook = self._run_response(flow)
-
-        extracted = flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]
-        expected_usage = _expected_usage(provider_case)
-        assert extracted["model"] == expected_usage["model"]
-        assert extracted["tokens.input"] == expected_usage["tokens.input"]
-        assert extracted["tokens.output"] == expected_usage["tokens.output"]
-        events = webhook.usage_events()
-        by_category = {event["category"]: event["quantity"] for event in events}
-        assert by_category == _expected_event_quantities(provider_case)
-
-    @pytest.mark.parametrize(
-        "provider_case",
-        MODEL_PROVIDER_JSON_CASES,
-        ids=_model_provider_json_case_id,
-    )
-    def test_full_pipeline_brotli_model_json_uses_bounded_fallback(
-        self, tmp_path, real_flow, provider_case
-    ):
-        """Brotli streaming decode is skipped, but bounded JSON fallback remains active."""
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            provider_case,
-            proxy_log_path=tmp_path / "proxy.jsonl",
-        )
-        payload = _standard_success_payload(provider_case)
-        compressed = brotli.compress(payload)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "application/json", "content-encoding": "br"}),
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(compressed)
-
-        webhook = self._run_response(flow)
-
-        extracted = flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]
-        expected_usage = _expected_usage(provider_case)
-        assert extracted["model"] == expected_usage["model"]
-        assert extracted["tokens.input"] == expected_usage["tokens.input"]
-        assert extracted["tokens.output"] == expected_usage["tokens.output"]
-        if provider_case.uses_openai_responses:
-            assert extracted["tokens.cache_read"] == expected_usage["tokens.cache_read"]
-        events = webhook.usage_events()
-        by_category = {event["category"]: event["quantity"] for event in events}
-        assert by_category == _expected_event_quantities(provider_case)
-
-    def test_full_pipeline_incomplete_model_json_does_not_report_partial_usage(
-        self, tmp_path, real_flow
-    ):
-        """Fields seen before EOF are ignored unless the JSON document completes."""
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            ANTHROPIC_JSON_CASE,
-            proxy_log_path=tmp_path / "proxy.jsonl",
-        )
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "application/json"}),
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(
-            b'{"id":"msg_1","model":"claude-sonnet-4-6",'
-            b'"usage":{"input_tokens":50,"output_tokens":200}'
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        proxy_log = Path(flow.metadata[metadata_keys.VM_PROXY_LOG_PATH])
-        entries = read_jsonl_entries_after_flush(proxy_log)
-        usage_warnings = [
-            entry
-            for entry in entries
-            if entry.get("message") == "Model provider JSON usage extraction failed"
-        ]
-        assert len(usage_warnings) == 1
-        assert usage_warnings[0]["error"] == "incomplete json"
-
-    def test_full_pipeline_corrupt_model_json_encoding_does_not_fallback_to_raw_buffer(
-        self, tmp_path, real_flow
-    ):
-        """A bad Content-Encoding must not parse raw stream_buffer and bill usage."""
-        raw_json = json.dumps(
-            {
-                "id": "msg_1",
-                "model": "claude-sonnet-4-6",
-                "usage": {"input_tokens": 50, "output_tokens": 200},
-            }
-        ).encode()
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            ANTHROPIC_JSON_CASE,
-            proxy_log_path=tmp_path / "proxy.jsonl",
-        )
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "application/json", "content-encoding": "gzip"}),
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(raw_json)
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        assert metadata_keys.MODEL_PROVIDER_USAGE not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER not in flow.metadata
-
-    def test_full_pipeline_model_json_ignores_usage_array_shape(self, tmp_path, real_flow):
-        """usage fields inside array elements must not be treated as usage object fields."""
-        body = json.dumps(
-            {
-                "id": "msg_1",
-                "model": "claude-sonnet-4-6",
-                "usage": [{"input_tokens": 50, "output_tokens": 200}],
-            }
-        ).encode()
-        flow = self._model_provider_flow(
-            real_flow,
-            tmp_path,
-            ANTHROPIC_JSON_CASE,
-            proxy_log_path=tmp_path / "proxy.jsonl",
-        )
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "application/json"}),
-        )
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(body)
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-
-    def test_no_usage_report_for_non_model_provider(self, tmp_path, real_flow):
-        """Non-model-provider requests should not trigger usage reporting."""
-        flow = real_flow(with_response=False, host="api.github.com")
-        proxy_log_path = tmp_path / "proxy.jsonl"
-        flow.metadata[metadata_keys.VM_RUN_ID] = "run-abc-123"
-        flow.metadata[metadata_keys.VM_NETWORK_LOG_PATH] = str(tmp_path / "network.jsonl")
-        flow.metadata[metadata_keys.VM_PROXY_LOG_PATH] = str(proxy_log_path)
-        flow.metadata[metadata_keys.FIREWALL_ACTION] = "ALLOW"
-        flow.metadata[metadata_keys.ORIGINAL_URL] = "https://api.github.com/repos"
-        flow.metadata[metadata_keys.FIREWALL_NAME] = "github"
-        body = b'{"incomplete":'
-        set_response_stream_buffer(flow, body)
-        flow.response = tutils.tresp(
-            status_code=200, headers=header_map({"content-type": "application/json"})
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        assert metadata_keys.MODEL_PROVIDER_USAGE not in flow.metadata
-        logging_utils.flush_log_path(str(proxy_log_path))
-        assert not proxy_log_path.exists()
-
-    @pytest.mark.parametrize("firewall_name", [None, 42])
-    def test_json_fallback_skips_malformed_firewall_name(self, tmp_path, real_flow, firewall_name):
-        flow = real_flow(with_response=False, host="api.anthropic.com")
-        self._set_common_model_metadata(flow, tmp_path)
-        flow.metadata[metadata_keys.ORIGINAL_URL] = ANTHROPIC_JSON_CASE.original_url
-        flow.metadata[metadata_keys.FIREWALL_NAME] = firewall_name
-        body = _standard_success_payload(ANTHROPIC_JSON_CASE)
-        set_response_stream_buffer(flow, body)
-        flow.response = tutils.tresp(
-            status_code=200, headers=header_map({"content-type": "application/json"})
-        )
-
-        webhook = self._run_response(flow)
-
-        assert webhook.request_count == 0
-        assert metadata_keys.MODEL_PROVIDER_USAGE not in flow.metadata
-
-
-class TestModelProviderResponseUsageWebhookDelivery:
-    """Tests for real background executor webhook delivery."""
-
-    def test_full_path_response_to_webhook(
-        self, tmp_path, real_flow, fresh_usage_executor, usage_webhook_api
-    ):
-        """Integration: response() -> _maybe_report -> _enqueue -> _retry -> webhook.
-
-        Verifies wiring between all intermediate layers through loopback HTTP.
-        """
-        flow = real_flow(with_response=False, host="api.anthropic.com")
-        log_path = str(tmp_path / "network.jsonl")
-        flow.metadata[metadata_keys.VM_RUN_ID] = "run-int-001"
-        flow.metadata[metadata_keys.VM_NETWORK_LOG_PATH] = log_path
-        flow.metadata[metadata_keys.FIREWALL_ACTION] = "ALLOW"
-        flow.metadata[metadata_keys.ORIGINAL_URL] = "https://api.anthropic.com/v1/messages"
-        flow.metadata[metadata_keys.FIREWALL_NAME] = "model-provider:anthropic-api-key"
-        flow.metadata[metadata_keys.FIREWALL_BILLABLE] = True
-        flow.metadata[metadata_keys.VM_SANDBOX_AUTH_KEY] = "tok-xyz"
-        flow.metadata[metadata_keys.MODEL_USAGE_PROVIDER] = "claude-sonnet-4-6"
-        flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE] = {
-            "model": "claude-sonnet-4-6",
-            "tokens.input": 100,
-            "tokens.output": 500,
-        }
-        flow.response = tutils.tresp(
-            status_code=200, headers=header_map({"content-type": "text/event-stream"})
-        )
-
-        with usage_webhook_api() as webhook:
-            mitm_addon.response(flow)
-            # Flush the executor to ensure the background POST completes
-            usage.flush_usage_events(trigger="test")
-            usage.webhook.usage_executor.shutdown(wait=True)
-
-        requests_by_path = {request.path: request for request in webhook.requests}
-        assert set(requests_by_path) == {
-            "/api/webhooks/agent/usage-event",
-            "/api/webhooks/agent/model-usage-observation",
-        }
-        body = requests_by_path["/api/webhooks/agent/usage-event"].json_body()
-        assert body["runId"] == "run-int-001"
-        by_category = {event["category"]: event for event in body["events"]}
-        assert by_category["tokens.input"]["quantity"] == 100
-        assert by_category["tokens.output"]["quantity"] == 500
-        assert by_category["tokens.input"]["provider"] == "claude-sonnet-4-6"
-        observation_body = requests_by_path[
-            "/api/webhooks/agent/model-usage-observation"
-        ].json_body()
-        observation_by_category = {event["category"]: event for event in observation_body["events"]}
-        assert observation_by_category["tokens.input"]["model"] == "claude-sonnet-4-6"
